refactor: split makepresentation editor files
