+++ product/akbun-makepresentation/workspace/src/index.html
@@ -554,11 +554,33 @@
     </div>
   </dialog>
 
+  <script src="editor/constants.js"></script>
+  <script src="editor/shapes.js"></script>
+  <script src="editor/geometry.js"></script>
+  <script src="editor/deck.js"></script>
+  <script src="editor/presets.js"></script>
+  <script src="editor/svg.js"></script>
+  <script src="editor/utils.js"></script>
   <script src="editor.js"></script>
   <script src="settings.js"></script>
   <script src="ai.js"></script>
   <script src="api.js"></script>
   <script src="ai-panel.js"></script>
+  <script src="renderer/render.js"></script>
+  <script src="renderer/history.js"></script>
+  <script src="renderer/canvas.js"></script>
+  <script src="renderer/text.js"></script>
+  <script src="renderer/presets.js"></script>
+  <script src="renderer/code.js"></script>
+  <script src="renderer/guidelines.js"></script>
+  <script src="renderer/slide-size.js"></script>
+  <script src="renderer/settings.js"></script>
+  <script src="renderer/clipboard.js"></script>
+  <script src="renderer/context-menu.js"></script>
+  <script src="renderer/properties.js"></script>
+  <script src="renderer/slides.js"></script>
+  <script src="renderer/files.js"></script>
+  <script src="renderer/presentation.js"></script>
   <script src="renderer.js"></script>
 </body>
 </html>
